+++ cSISTEMICA.h
@@ -8,13 +8,13 @@ class cSISTEMICA :
     public cRADIOTERAPIA
 {
 public:
-    static unsigned int dosisMaxPaciente;
-    static unsigned int dosisMaxTumor;
+    static float dosisMaxPaciente;
+    static float dosisMaxTumor;
 
 
     cSISTEMICA(eTamanioTumor tamanio);
     ~cSISTEMICA();
-    unsigned int ajustar_dosis(eTamanioTumor tamanio);
+    float ajustar_dosis(eTamanioTumor tamanio);
     string to_string();
 };
